requirements/local.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.16
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-model-utils==5.0.0
Pillow==7.0.0
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.1
typing-extensions==4.12.2
tzdata==2024.2
Unipath==1.1